Hopefully Final Interpreter
Fixed the scope errors described in our last Interpreter submission feedback. This involved changes to visiting function, if statement, and switch statement. Added new scope creation to switch/if. Changed function so that when a function is called, the scope that it was defined with is used, not the scope it is called with.